Move test for checking test key usage.
